A basic implementation that works.
