+++ pretty_poly/__init__.py
@@ -1,2 +1,1 @@
 from .ascii import make_ascii
-from .png import make_png
