+++ tests/test_png.py
@@ -1,4 +1,8 @@
-from pretty_poly import make_png
+import tempfile
+
+import png
+
+from pretty_poly.png import write_colored_blocks_png
 
 
 def test_basic_png_right_size():
@@ -6,6 +10,10 @@ def test_basic_png_right_size():
         [(0, 1), (0, 0)],
         [(1, 1), (1, 0)]
     ]
-    image = make_png(tiling)
-    assert image.width == 20
-    assert image.height == 20
+    f, filename = tempfile.mkstemp(suffix=".png")
+    write_colored_blocks_png(filename, tiling)
+    r = png.Reader(filename)
+    width, height, rows, info = r.read()
+    assert width == 20
+    assert height == 20
+
